Move lexer to other module
