+++ Lexer/src/main/java/mipt/compiler/minirust/lexer/tokens/IdentifierToken.java
@@ -1,3 +1,3 @@
-package mipt.compiler.minirust.parser.tokens;
+package mipt.compiler.minirust.lexer.tokens;
 
 public record IdentifierToken(String key) implements TokenType {}
