+++ Lexer/src/main/java/mipt/compiler/minirust/lexer/tokens/TokenType.java
@@ -1,4 +1,4 @@
-package mipt.compiler.minirust.parser.tokens;
+package mipt.compiler.minirust.lexer.tokens;
 
 public sealed interface TokenType
     permits DelimitersToken, IdentifierToken, IntegerLiteral, KeywordsToken, PunctuationsToken {}
