+++ settings.gradle.kts
@@ -1,2 +1,3 @@
 rootProject.name = "MiniRustCompiler"
 include("Parser")
+include("Lexer")
